MCP-Playground/pyproject.toml
[project]
name = "mcp-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.34.2",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "langgraph-prebuilt>=0.2.2",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
